python/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
certifi==2025.1.31
    # via requests
charset-normalizer==3.4.1
    # via requests
idna==3.10
    # via requests
python-dotenv==1.1.0
    # via -r requirements.in
pyyaml==6.0.2
    # via -r requirements.in
requests==2.32.3
    # via -r requirements.in
urllib3==2.4.0
    # via requests
